Added chainable db function calls in DBManager. Separated business logic from DBManager into respective 'entities'.
